edit type of id in migrations

+++ database/migrations/2022_07_13_105740_create_adrresses_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('addresses', function (Blueprint $table) {
             $table->id();
-            $table->unsignedSmallInteger('city_id');
+            $table->unsignedBigInteger('city_id');
             $table->string('shipping_method');
             $table->integer('price');
             $table->string('address');
